Layer handling reworked
Instead of having a primary and a temporary layer, store the state of at
most 32 layers in a bitfield. These can be individually turned on and
off, and key lookup starts from the top, and goes downwards until the
default layer to find a non-transparent key.

This allows one to reuse a partially transparent layer: set the default,
and the transparent parts will be reused. The numpad layer was updated
accordingly.

Having an interface to the layer switching things also makes it easier
to build other behaviour on top of these.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>