Layer_::lookup: Rearrange and fix the loop
Instead of a do {} while loop, put the condition first, and fix it, so
that it actually tests the thing we want.

The loop should have started from the topmost layer, check if its
active, if so, look up the key, and go a layer down. Then continue this
until it finds a non-transparent key, or reaches below the default
layer. Instead, it stopped on transparent keys.

Rearranging the loop helped make the intent clearer, and the code now
matches that.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>